#1352 Reference data wizards in monitor: css styling of input fields (max width); checking unique name; "Reference data" item in main menu; Regex swap string pattern wizard -- correction of page indexes.

+++ monitor/extensions/referencedata-wizards/src/test/java/org/datacleaner/monitor/server/wizard/stringpattern/regexswap/RegexSwapStringPatternReferenceDataPage2Test.java
@@ -29,7 +29,7 @@ public class RegexSwapStringPatternReferenceDataPage2Test {
     @Test
     public void testGetPageIndex() throws Exception {
         RegexSwapStringPatternReferenceDataPage2 page = new RegexSwapStringPatternReferenceDataPage2(null);
-        assertEquals(new Integer(0), page.getPageIndex());
+        assertEquals(new Integer(1), page.getPageIndex());
     }
 
     @Test
